requirements/base.txt
fastapi
gino
email-validator
passlib[bcrypt]
pyjwt
pyyaml
uvicorn
alembic
psycopg2-binary